Modularization of the FormComponet component. + creating ButtonLogout.tsx
