+++ frontend/src/components/FormComponent/FormComponent.jsx
@@ -1,6 +1,7 @@
 //FormComponent.jsx
 import { useState } from "react";
 import PropTypes from "prop-types";
+import { ButtonLogout } from "../ButtonLogout/ButtonLogout";
 
 const FormComponent = ({ onSubmit, username, onLogout }) => {
   const [message, setMessage] = useState("");
@@ -11,10 +12,6 @@ const FormComponent = ({ onSubmit, username, onLogout }) => {
     setMessage("");
   };
 
-  const handleLogout = () => {
-    onLogout();
-  };
-
   return (
     <div className="max-w-md w-full">
       {username && (
@@ -23,23 +20,7 @@ const FormComponent = ({ onSubmit, username, onLogout }) => {
             Welcome, {username}!
           </span>
 
-          <svg
-            className="h-8 w-8 text-red-500 hover:text-red-700"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            strokeWidth="2"
-            stroke="currentColor"
-            fill="none"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            onClick={handleLogout}
-          >
-            <title>Logout</title>
-            <path stroke="none" d="M0 0h24v24H0z" />
-            <path d="M14 8v-2a2 2 0 0 0 -2 -2h-7a2 2 0 0 0 -2 2v12a2 2 0 0 0 2 2h7a2 2 0 0 0 2 -2v-2" />
-            <path d="M7 12h14l-3 -3m0 6l3 -3" />
-          </svg>
+          <ButtonLogout onLogout={onLogout} />
         </div>
       )}
